# backend/cmd/api/docs/swagger.yaml
definitions:
...
  models.InventoryUpdate:
    properties:
      product_id:
        type: integer
      stock:
        type: integer
    type: object
...
  models.OfferMaking:
    properties:
      category_id:
        type: integer
      discount:
        type: integer
    type: object
...
  models.Search:
    properties:
      searchkey:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /admin/inventories:
    delete:
      consumes:
      - application/json
      description: Admin can delete a product
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete Inventory
      tags:
      - Admin
    post:
      consumes:
      - multipart/form-data
      description: Admin can add new  products
      parameters:
      - description: category_id
        in: formData
        name: category_id
        required: true
        type: string
      - description: product_name
        in: formData
        name: product_name
        required: true
        type: string
      - description: size
        in: formData
        name: size
        required: true
        type: string
      - description: price
        in: formData
        name: price
        required: true
        type: string
      - description: stock
        in: formData
        name: stock
        required: true
        type: string
      - description: image
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Add Inventory
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Admin can update stock of the inventories
      parameters:
      - description: update stock
        in: body
        name: add-stock
        required: true
        schema:
          $ref: '#/definitions/models.InventoryUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update Stock
      tags:
      - Admin
  /admin/offers/add:
    post:
      consumes:
      - application/json
      description: Admin can add new offers forspecified categories
      parameters:
      - description: offer
        in: body
        name: offer
        required: true
        schema:
          $ref: '#/definitions/models.OfferMaking'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Add Offer
      tags:
      - Admin
...
  /users/home/products:
    get:
      consumes:
      - application/json
      description: user can view the list of available products
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: List Products
      tags:
      - User
  /users/home/products/details:
    get:
      consumes:
      - application/json
      description: user can view the details of the product
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Show Product Details
      tags:
      - User
...
  /users/search:
    post:
      consumes:
      - application/json
      description: user can search with a key and get the list of  products similar
        to that key
      parameters:
      - description: search
        in: body
        name: key
        required: true
        schema:
          $ref: '#/definitions/models.Search'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Search Products
      tags:
      - User
...
swagger: "2.0"